requirements.txt
flask
werkzeug
jinja2
itsdangerous
click
pillow
gunicorn
psycopg2-binary
